Suggest a package name based on a development package title or description

If the package you are using already has a title, simply pass the path to the package root in path. Otherwise use title to specify a potential title.

Usage

suggest(path = ".", field = c("Title", "Description"), text = NULL)

Arguments

path

Path to a existing package to extract the title from.

field

one of "Title" or "Description"

text

text string to search.

Examples

if (FALSE) { # \dontrun{
# Default will use the title from the current path.
suggest()

# Can also suggest based on the description
suggest(field = "Description")
} # }

# Or by explicitly using the text argument
suggest(
  text =
    "A Package for Displaying Visual Scenes as They May Appear to an Animal with Lower Acuity"
)
#> displayingrvis